Update install.sh
Detect RH.

+++ install.sh
@@ -22,6 +22,15 @@ ln -s $DOTFILES/screenrc ~/.screenrc
 mv ~/.dir_colors ~/.dir_colors.old 2> /dev/null
 ln -s $DOTFILES/dircolors/dircolors.ansi-dark ~/.dir_colors
 
-sudo apt-get install exuberant-ctags vim --yes
+if [ -f /etc/redhat-release ]
+then
+    sudo yum -y install ctags-etags
+fi
+
+if [ -f /etc/debian_version ]
+then
+    sudo apt-get install exuberant-ctags vim --yes
+fi
+
 git submodule update --init
